# services/web-app/openapi.yaml
openapi: 3.0.3
info:
  title: RespiraAlly API
  description: API for the RespiraAlly respiratory care management system
  version: 1.0.0
  contact:
    name: RespiraAlly Team
    email: support@respiraally.com
servers:
  - url: http://localhost:5000/api/v1
    description: Development server
  - url: https://api.respiraally.com/api/v1
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: "INVALID_INPUT"
            message:
              type: string
              example: "Request body cannot be empty"
...
    Patient:
      type: object
      properties:
        user_id:
          type: integer
          example: 123
        first_name:
          type: string
          example: "美麗"
        last_name:
          type: string
          example: "陳"
        gender:
          type: string
          example: "female"
        email:
          type: string
          example: "meili.chen@example.com"
        phone:
          type: string
          example: "0987654321"
        health_profile:
          $ref: '#/components/schemas/HealthProfile'
    HealthProfile:
      type: object
      properties:
        height_cm:
          type: integer
          example: 160
        weight_kg:
          type: integer
          example: 55
        smoke_status:
          type: string
          enum: [never, former, current]
          example: "never"
        updated_at:
          type: string
          format: date-time
          example: "2025-01-15T10:30:00Z"
...
    PaginationMeta:
      type: object
      properties:
        total_items:
          type: integer
          example: 100
        total_pages:
          type: integer
          example: 10
        current_page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 10
paths:
...
  /therapist/patients:
    get:
      summary: "獲取管理的病患列表"
      description: "獲取當前登入的呼吸治療師所管理的所有病患的簡要列表，支援風險篩選和分頁"
      tags: ["Patients"]
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: "頁碼"
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
          description: "每頁數量"
        - name: risk
          in: query
          schema:
            type: string
            enum: [high, medium, low]
          description: "風險等級篩選"
        - name: limit
          in: query
          schema:
            type: integer
          description: "限制返回數量（覆蓋per_page）"
        - name: sort_by
          in: query
          schema:
            type: string
            default: "last_login"
          description: "排序欄位"
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: "desc"
          description: "排序順序"
      responses:
        '200':
          description: "成功獲取病患列表"
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Patient'
                  pagination:
                    $ref: '#/components/schemas/PaginationMeta'
        '400':
          description: "參數驗證失敗"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: "Token 無效或未提供"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: "沒有治療師權限"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /patients/{patient_id}/profile:
    get:
      summary: "獲取病患詳細健康檔案"
      description: "獲取指定病患的詳細健康檔案資訊"
      tags: ["Patients"]
      security:
        - bearerAuth: []
      parameters:
        - name: patient_id
          in: path
          required: true
          schema:
            type: integer
          description: "病患的 user_id"
      responses:
        '200':
          description: "成功獲取病患檔案"
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Patient'
        '401':
          description: "Token 無效或未提供"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: "沒有權限查看此病患"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: "找不到該病患"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /patients/{patient_id}/kpis:
    get:
      summary: "獲取個別病患 KPI 指標"
      description: "獲取指定病患的關鍵績效指標，包括最新問卷分數、依從性統計等"
      tags: ["Patients"]
      security:
        - bearerAuth: []
      parameters:
        - name: patient_id
          in: path
          required: true
          schema:
            type: integer
          description: "病患的 user_id"
        - name: days
          in: query
          schema:
            type: integer
            default: 7
          description: "計算天數範圍（默認7天）"
      responses:
        '200':
          description: "成功獲取病患 KPI"
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    description: "KPI 資料"
                  meta:
                    type: object
                    properties:
                      patient_id:
                        type: integer
                      calculation_days:
                        type: integer
                      calculated_at:
                        type: string
                        format: date-time
        '401':
          description: "Token 無效或未提供"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: "沒有權限查看此病患"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: "找不到該病患"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Questionnaire endpoints
...